make example for android
