+++ examples/Basic/Basic.js
@@ -1,0 +1,128 @@
+const React = require('react-native');
+const SideMenu = require('react-native-side-menu');
+const Menu = require('./Menu');
+
+const {
+  AppRegistry,
+  StyleSheet,
+  Text,
+  View,
+  Image,
+  ScrollView,
+  TouchableOpacity,
+  Component
+} = React;
+
+const styles = StyleSheet.create({
+  button: {
+    position: 'absolute',
+    bottom: 50,
+    backgroundColor: 'red',
+    borderRadius: 20,
+  },
+  button2: {
+    position: 'absolute',
+    bottom: 20,
+    backgroundColor: 'red',
+    borderRadius: 20,
+  },
+  caption: {
+    fontSize: 20,
+    fontWeight: 'bold',
+    alignItems: 'center',
+  },
+  container: {
+    flex: 1,
+    justifyContent: 'center',
+    alignItems: 'center',
+    backgroundColor: '#F5FCFF',
+  },
+  welcome: {
+    fontSize: 20,
+    textAlign: 'center',
+    margin: 10,
+  },
+  instructions: {
+    textAlign: 'center',
+    color: '#333333',
+    marginBottom: 5,
+  },
+});
+
+class Button extends Component {
+  handlePress(e) {
+    this.context.menuActions.toggle();
+    if (this.props.onPress) {
+      this.props.onPress(e);
+    }
+  }
+
+  render() {
+    return (
+      <TouchableOpacity
+        onPress={this.handlePress.bind(this)}
+        style={this.props.style}>
+        <Text>{this.props.children}</Text>
+      </TouchableOpacity>
+    );
+  }
+}
+
+Button.contextTypes = {
+  menuActions: React.PropTypes.object.isRequired
+};
+
+class Basic extends Component {
+  constructor(props, ctx) {
+    super(props, ctx);
+
+    this.state = {
+      touchToClose: false,
+    };
+  }
+
+  handleOpenWithTouchToClose() {
+    this.setState({
+      touchToClose: true,
+    });
+  }
+
+  handleChange(isOpen) {
+    if (!isOpen) {
+      this.setState({
+        touchToClose: false,
+      });
+    }
+  }
+
+  render() {
+    return (
+      <SideMenu
+        menu={<Menu />}
+        touchToClose={this.state.touchToClose}
+        onChange={this.handleChange.bind(this)}>
+        <View style={styles.container}>
+          <Text style={styles.welcome}>
+            Welcome to React Native!
+          </Text>
+          <Text style={styles.instructions}>
+            To get started, edit index.ios.js
+          </Text>
+          <Text style={styles.instructions}>
+            Press Cmd+R to reload,{'\n'}
+            Cmd+Control+Z for dev menu
+          </Text>
+        </View>
+        <Button style={styles.button}>
+          Toggle menu
+        </Button>
+        <Button style={styles.button2}
+          onPress={this.handleOpenWithTouchToClose.bind(this)}>
+          Open menu (Overlay Closes)
+        </Button>
+      </SideMenu>
+    );
+  }
+}
+
+module.exports = Basic;
